requirements.txt
beautifulsoup4==4.12.3
dateparser==1.2.0
python-dotenv==1.0.1
Requests==2.31.0
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
seaborn==0.13.2
bar-chart-race==0.1.0
python-dotenv==1.0.1
